Initial class methods written
Garoupa.make_groups will make groups from a list (an Array) and pass the groups on to Garoupa.new to instantiate a new Garoupa instance.  The method allows for some optional parameters:  target groups size, max difference in group sizes in case the groups aren't even, and history of past groups so that groups can be mad with a preference for placing list items with new groupmates.

Garoupa.past_groupmates takes a list and previous group history and returns a Hash object.  Keys are the list items.  Values are collections of list items that have been groupmates of the key.